requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.1
django-admin-rangefilter==0.9.0
django-bootstrap-modal-forms==2.2.0
django-filter==22.1
django-phone-field==1.8.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
pytils==0.4.1
pytz==2022.2.1
requests==2.28.1
selenium==3.14.0
sqlparse==0.4.3
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
